### 2022-02-23 #4194 @editorialbot
[PRE REVIEW]: Connectome Mapper 3: A Flexible and Open-Source Pipeline Software for Multiscale Multimodal Human Connectome Mapping
**Submitting author:** @sebastientourbier (<a href="http://orcid.org/0000-0002-4441-899X">Sebastien Tourbier</a>)
**Repository:** https://github.com/connectomicslab/connectomemapper3
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.3
