### 2022-02-23 #4194 @arfon
@sebastientourbier, for authors with multiple affiliations, please quote them. e.g.:

```
  - name: Meritxell Bach Cuadra
    affiliation: "7,1"
    orcid: 0000-0003-2730-4285
```
