+++ paper/paper.bib
@@ -266,7 +266,7 @@
     pages = {825--841},
     volume = {17},
     url = {http://linkinghub.elsevier.com/retrieve/pii/S1053811902911328},
-    doi = {10.1016/S1053-8119(02)91132-8},
+    doi = {{10.1016/S1053-8119%2802%2991132-8}},
     issn = {10538119},
     pmid = {12377157},
     keywords = {Accuracy, Affine transformation, Global optimization, Motion correction, Multimodal registration, Multiresolution search, Robustness}
@@ -308,7 +308,7 @@
     number = {3},
     pages = {353--364},
     volume = {20},
-    doi = {10.1038/nn.4502.Network}
+    doi = {10.1038/nn.4502}
 }
 
 @book{Colon-perez2018,
@@ -785,7 +785,7 @@
     author = {Delettre, Celine and Dell, Leigh-anne and Delettre, Céline and Mess{\'{e}}, Arnaud and Dell, Leigh-anne and Foubet, Ophélie and Heuer, Katja and Larrat, Benoit and Meriaux, Sebastien and Mangin, Jean-francois and Reillo, Isabel and Romero, Camino De Juan and Borrell, Victor and Toro, Roberto and Hilgetag, Claus C},
     number = {May},
     url = {http://dx.doi.org/10.1162/netn_a_00098},
-    doi = {10.1162/netn},
+    doi = {10.1162/netn_a_00098},
     keywords = {Validation,Diffusion MRI,Tractography,Tract-tracin}
 }
 
@@ -1102,7 +1102,7 @@
     pages = {774--781},
     volume = {62},
     url = {http://www.sciencedirect.com/science/article/pii/S1053811912000389},
-    doi = {https://doi.org/10.1016/j.neuroimage.2012.01.021},
+    doi = {10.1016/j.neuroimage.2012.01.021},
     issn = {1053-8119},
     keywords = {MRI, Morphometry, Registration, Segmentation}
 }
@@ -1465,7 +1465,7 @@
     month = {1},
     pages = {551739},
     url = {http://biorxiv.org/content/early/2019/02/15/551739.abstract},
-    doi = {https://doi.org/10.1101/551739}
+    doi = {10.1101/551739}
 }
 
 @article{Tournier2012MRtrix:Regions,
@@ -1491,7 +1491,7 @@
     pages = {51--62},
     volume = {12},
     url = {http://doi.wiley.com/10.1002/%28SICI%291099-1492%28199902%2912%3A1%3C51%3A%3AAID-NBM546%3E3.0.CO%3B2-E},
-    doi = {10.1002/(SICI)1099-1492(199902)12:1<51::AID-NBM546>3.0.CO;2-E},
+    doi = {10.1002/%28SICI%291099-1492%28199902%2912%3A1%3C51%3A%3AAID-NBM546%3E3.0.CO%3B2-E},
     issn = {0952-3480},
     keywords = {Apparent diffusion coefficient (ADC), Biexponential b-factor decay, Brain}
 }
@@ -1792,7 +1792,7 @@
 }
 
 @article{Peng2011ReproducibleScience,
-    title = {{Reproducible Research in Computing Science}},
+    title = {{Reproducible Research in Computional Science}},
     year = {2011},
     journal = {Science},
     author = {Peng, Roger D},
@@ -1800,7 +1800,7 @@
     pages = {1226--1227},
     volume = {334},
     isbn = {1095-9203 (Electronic){\textbackslash}r0036-8075 (Linking)},
-    doi = {10.1126/science.1213847.Reproducible},
+    doi = {10.1126/science.1213847},
     issn = {1095-9203},
     pmid = {22144613},
     arxivId = {0901.4552}
[…]
