### 2022-02-24 4e2c73b
REV: Clean paper.bib file [skip ci]
