+++ paper/paper.bib
@@ -1,58 +1,12 @@
-@software{ZenodoCMP:2021,
-  author       = {Tourbier, Sebastien and
-                  Aleman-Gomez, Yasser and
-                  Mullier, Emeline and
-                  Griffa, Alessandra and
-                  Bach Cuadra, Meritxell and
-                  Hagmann, Patric},
-  title        = {{connectomicslab/connectomemapper3: Connectome
-                   Mapper v3.0.0}},
-  month        = feb,
-  year         = 2022,
-  publisher    = {Zenodo},
-  version      = {v3.0.3},
-  doi          = {10.5281/zenodo.3475969},
-  url          = {https://doi.org/10.5281/zenodo.3475969}
-}
-
-@article{Greve2009,
-    title = {{Accurate and robust brain image alignment using boundary-based registration}},
-    year = {2009},
-    journal = {NeuroImage},
-    author = {Greve, Douglas N. and Fischl, Bruce},
-    number = {1},
-    month = {10},
-    pages = {63--72},
-    volume = {48},
-    url = {https://linkinghub.elsevier.com/retrieve/pii/S1053811909006752},
-    doi = {10.1016/j.neuroimage.2009.06.060},
-    issn = {10538119},
-    pmid = {19573611}
-}
-
-@article{Kiar2017,
-    title = {{AComprehensive Cloud Framework for Scalable, Reliable, and Replicable Human Connectome Es- timation and Meganalysis}},
-    year = {2018},
-    journal = {bioRxiv},
-    author = {Kiar, Gregory and Bridgeford, Eric and Roncal, William Gray and (CoRR), Consortium for Reliability and Reproducibliity and Chandrashekhar, Vikram and Mhembere, Disa and Ryman, Sephira and Zuo, Xi-Nian and Marguiles, Daniel S and Craddock, R Cameron and Priebe, Carey E and Jung, Rex and Calhoun, Vince and Caffo, Brian and Burns, Randal and Milham, Michael P and Vogelstein, Joshua},
-    pages = {188706},
-    url = {https://www.biorxiv.org/content/early/2017/09/16/188706 https://www.biorxiv.org/content/10.1101/188706v6},
-    doi = {10.1101/188706}
-}
-
-@article{Cox1996,
-    title = {{AFNI: Software for Analysis and Visualization of Functional Magnetic Resonance Neuroimages}},
-    year = {1996},
-    journal = {Computers and Biomedical Research},
-    author = {Cox, Robert W.},
-    number = {3},
-    month = {6},
-    pages = {162--173},
-    volume = {29},
-    url = {https://linkinghub.elsevier.com/retrieve/pii/S0010480996900142},
-    doi = {10.1006/cbmr.1996.0014},
-    issn = {00104809},
-    pmid = {8812068}
+@software{ZenodoCMP:2022,
+    author = {Tourbier, Sebastien and Aleman-Gomez, Yasser and Mullier, Emeline and Griffa, Alessandra and Bach Cuadra, Meritxell and Hagmann, Patric},
+    title = {{connectomicslab/connectomemapper3: Connectome Mapper v3.0.3}},
+    month = {2},
+    year = {2022},
+    publisher = {Zenodo},
+    version = {v3.0.3},
+    doi = {10.5281/zenodo.3475969},
+    url = {https://doi.org/10.5281/zenodo.3475969}
 }
 
 @article{Cox2012,
@@ -70,29 +24,6 @@
     issn = {1053-8119}
 }
 
-@article{Zuo2014,
-    title = {{An open science resource for establishing reliability and reproducibility in functional connectomics}},
-    year = {2014},
-    journal = {Scientific Data},
[…]
