+++ paper/paper.bib
@@ -686,3 +686,45 @@
     version = {0.2.0},
     url = {https://github.com/mne-tools/mne-connectivity},
 }
+
+@article{Zheng2020GeometricConnectomeb,
+    title = {{Geometric renormalization unravels self-similarity of the multiscale human connectome}},
+    year = {2020},
+    journal = {Proceedings of the National Academy of Sciences},
+    author = {Zheng, Muhua and Allard, Antoine and Hagmann, Patric and Alem{\'{a}}n-G{\'{o}}mez, Yasser and Serrano, M. Ángeles},
+    number = {33},
+    month = {8},
+    pages = {20244--20253},
+    volume = {117},
+    publisher = {National Academy of Sciences},
+    url = {https://www.pnas.org/content/117/33/20244},
+    doi = {10.1073/PNAS.1922248117},
+    issn = {0027-8424},
+    pmid = {32759211}
+}
+
+@
+
+@misc{BellecNeuroimagingNIAK:2016,
+    title = {{Neuroimaging Analysis Kit - Progress and challenges for standardized fMRI processing}},
+    month = {10},
+    year = {2016},
+    author = {Bellec, Pierre},
+    howpublished = "\url{https://www.mcgill.ca/bic/files/bic/2016_10-17_-_neuroimage_analysis_kit_-_pierre_bellec.pdf}",
+}
+
+@article{AVANTS2008SymmetricBrain,
+    title = {{Symmetric diffeomorphic image registration with cross-correlation: Evaluating automated labeling of elderly and neurodegenerative brain}},
+    year = {2008},
+    journal = {Medical Image Analysis},
+    author = {Avants, B. and Epstein, C. and Grossman, M. and GEE, J.},
+    number = {1},
+    month = {2},
+    pages = {26--41},
+    volume = {12},
+    url = {https://linkinghub.elsevier.com/retrieve/pii/S0022202X15370834 https://linkinghub.elsevier.com/retrieve/pii/S1361841507000606},
+    doi = {10.1016/j.media.2007.06.004},
+    issn = {13618415},
+    arxivId = {NIHMS150003},
+    keywords = {epiblast, gfp fusion, histone h2b-, icm, lineage specification, live imaging, mouse blastocyst, pdgfr {$\alpha$}, primitive endoderm}
+}
