### 2022-02-24 c8fa689
FIX: Correct typo in paper.bib [skip ci]
