+++ paper/paper.bib
@@ -703,8 +703,6 @@
     pmid = {32759211}
 }
 
-@
-
 @misc{BellecNeuroimagingNIAK:2016,
     title = {{Neuroimaging Analysis Kit - Progress and challenges for standardized fMRI processing}},
     month = {10},
