### 2022-02-24 3a1f554
REV: Correct 2 DOIs not working in proof reading version [skip ci]
