+++ paper/paper.bib
@@ -231,7 +231,7 @@
     number = {2},
     pages = {125--127},
     volume = {2},
-    doi = {10.1162/netn{\_}e{\_}00052}
+    doi = {10.1162/netn_e_00052}
 }
 
 @article{Fischl2012FreeSurfer,
@@ -255,7 +255,7 @@
     volume = {3230},
     school = {EPFL},
     isbn = {9783642312083},
-    doi = {10.1007/978-3-642-31208-3{\_}6}
+    doi = {10.5075/epfl-thesis-3230}
 }
 
 @article{Jenkinson2012FSL,
