### 2022-02-24 #4194 @sebastientourbier
@arfon I reviewed my BibTex file and only kept the entries that are used in the paper. I also took care of adding a few missing DOI and correcting a few problematic ones. Now it seems everything is fine with `check references` and by checking the generated PDF.
