### 2022-04-07 connectomicslab/connectomemapper3#154 @jsheunis
LICENSE / COPYRIGHT file content
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

Hi @sebastientourbier

Based on the JOSS review checklist and [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), the repo has to contain a LICENSE or COPYING file with content that is [OSI approved](https://opensource.org/licenses/alphabetical). At first glance, your COPYRIGHT file doesn't seem to conform to these criteria, even though it is clear in its intent. I am not sure how closely the JOSS editors expect the criteria to be followed, so perhaps it's a good idea to check with them. Something else worth checking is whether putting restrictions on the use of the code, i.e.
> THIS SOFTWARE IS FOR RESEARCH PURPOSES ONLY AND SHALL NOT BE USED FOR
ANY CLINICAL USE.

is actually aligned with OSI approved licenses (I am not sure about this.)
