### 2022-04-07 connectomicslab/connectomemapper3#154 @jsheunis
And I see individual modules mention:

> This software is distributed under the open-source license Modified BSD.

although Modified BSD isn't mentioned in the COPYRIGHT file.
