### 2022-04-07 connectomicslab/connectomemapper3#154 @jsheunis
And [the docs license](https://connectome-mapper-3.readthedocs.io/en/latest/LICENSE.html) has the heading *BSD 3-clause license*. So I think it would be useful to make the different entries referring to license/copyright as consistent as possible.
