### 2022-04-07 connectomicslab/connectomemapper3#156 @jsheunis
DOC: Add size of docker image to docs
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

It could be useful to add a sentence to the Installation instructions that mention the size of the docker image that will be pulled, so that users know what to expect (6.27GB compressed for v3.0.3, according to https://hub.docker.com/r/sebastientourbier/connectomemapper-bidsapp/tags)
