### 2022-04-07 connectomicslab/connectomemapper3#157 @jsheunis
DOC: possibly use `sphinx-copybutton`
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

With [`sphinx-copybutton`](https://sphinx-copybutton.readthedocs.io/en/latest/) you can automatically add a button to all code blocks in the documentation that copies the code to the clipboard. This makes for an easier user experience when using code from the docs, but isn't an essential component.
