### 2022-04-07 connectomicslab/connectomemapper3#158 @jsheunis
DOC: conda environment creation instructions inconsistent
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

The README instructs the user to create a conda environment as follows:

```
conda create env -f /path/to/environment[_macosx].yml
```
 which fails, since it should be as provided in the docs:

```
conda env create -f /path/to/environment[_macosx].yml
```

An alternative would be:
```
conda create --file /path/to/environment[_macosx].yml
```
