### 2022-04-08 connectomicslab/connectomemapper3#159 @jsheunis
DOC: refer to contributing guidelines in README
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

I really like the detailed [contributing guidelines](https://connectome-mapper-3.readthedocs.io/en/latest/contributing.html), especially the distinction of types of contributions, and guide on how to install developer packages, create a branch, PR, etc.

To make this easier to discover, I think it would be good to refer to it in the repo README as well, perhaps as a short paragraph in the current [Contributors](https://github.com/connectomicslab/connectomemapper3#contributors-) section, or a separate Contributing section.
