### 2022-04-08 connectomicslab/connectomemapper3#160 @jsheunis
ModuleNotFoundError: No module named 'cmp.stages.eeg'
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

I get the following error when running `cmpbidsappmanager` from the conda environment `py37cmp-gui` that was created with `conda env create -f environment_macosx.yml` (and after creation running `pip install connectomemapper`). This prevents the GUI from opening.

```
(py37cmp-gui) ➜  ~ cmpbidsappmanager
Traceback (most recent call last):
  File "/Users/jsheunis/opt/miniconda3/envs/py37cmp-gui/bin/cmpbidsappmanager", line 5, in <module>
    from cmp.cli.cmpbidsappmanager import main
  File "/Users/jsheunis/opt/miniconda3/envs/py37cmp-gui/lib/python3.7/site-packages/cmp/cli/cmpbidsappmanager.py", line 21, in <module>
    import cmp.bidsappmanager.gui.principal
  File "/Users/jsheunis/opt/miniconda3/envs/py37cmp-gui/lib/python3.7/site-packages/cmp/bidsappmanager/gui/principal.py", line 21, in <module>
    import cmp.project
  File "/Users/jsheunis/opt/miniconda3/envs/py37cmp-gui/lib/python3.7/site-packages/cmp/project.py", line 25, in <module>
    from cmp.pipelines.functional import eeg as EEG_pipeline
  File "/Users/jsheunis/opt/miniconda3/envs/py37cmp-gui/lib/python3.7/site-packages/cmp/pipelines/functional/eeg.py", line 17, in <module>
    from cmp.stages.eeg.inverse_solution import EEGInverseSolutionStage
ModuleNotFoundError: No module named 'cmp.stages.eeg'
```

More info:
- macOS Big Sur 11.2.3
- conda 4.9.2
- Python 3.7.12
