### 2022-04-08 connectomicslab/connectomemapper3#161 @jsheunis
JOSS paper comments
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

The current version of the [JOSS paper](https://github.com/jsheunis/connectomemapper3/blob/master/paper/paper.md) is well written and clearly explains the need and functionality of CMP3. I have a few comments:
- While reading through the paper content, I adapted the wording slightly in a few places. These are (in my opinion) minor improvements, and you can use your discretion to decide whether or not to include these changes. I'll submit a small PR shortly.
- The JOSS guidelines state that the paper should have _"A summary describing the high-level functionality and purpose of the software for a diverse, non-specialist audience"_. At the moment, the summary section makes out the bulk of the paper, and I consider much of it to be quite detailed and specialist. I don't think this content is a problem though, just that I'm missing a shorter summary for a *diverse, non-specialist audience*. Perhaps a new paragraph could be included to serve as such a summary, while the current **Summary** section could be renamed to **Functionality** (or something similar). That is if you (and the JOSS editors) agree that this is indeed a useful approach.
- The **Mention** section's title seems a bit out of context, perhaps rename it to **Software use** or **Ongoing research** or **Community impact**, or something along those lines.
