### 2022-04-08 #4248 @jsheunis
@osorensen @sebastientourbier

Review update: I've checked most of the boxes in my checklist. The ones that are still open are dependent on open issues, particularly:
- https://github.com/connectomicslab/connectomemapper3/issues/154
- https://github.com/connectomicslab/connectomemapper3/issues/160
- https://github.com/connectomicslab/connectomemapper3/issues/161 (second point)
