### 2022-04-08 connectomicslab/connectomemapper3#155 @jsheunis
Perhaps the expected time it will take to download could also be useful? Or just a note that this process might take some time. It took around 40min to pull the image on my Macbook pro with a decent internet connection speed.

Likelwise, creating the conda environment also took very long.
