### 2022-04-08 ee538e0
DOC: Update docs/LICENSE.rst accordingly to to LICENSE
