### 2022-04-08 1c720b3
DOC: Correct `conda env create` instruction in README
