### 2022-04-09 #4248 @sebastientourbier
Dear @jsheunis,

Thank you a lot for the time you spent reviewing CMP3 and its paper!
I am going to address your issues next week and I will reply directly inside them for the sake of clarity.
I will let you all know as soon as it is done.

Best regards
