### 2022-04-11 connectomicslab/connectomemapper3#160 @sebastientourbier
Dear @jsheunis,

I am sorry to hear you encountered this issue.

Indeed, this has been recently reported in #150  and the cause has been identified.

Briefly, CMP3 is using two different setup.py files:

-  https://github.com/connectomicslab/connectomemapper3/blob/master/setup.py installs CMP3 with all data resources (such as atlas images). It is used during development and for installing CMP3 in the BIDS App (docker image).
-  https://github.com/connectomicslab/connectomemapper3/blob/master/setup_pypi.py is a lightweight installation of CMP3 (no data resources) that allows us to make python package < 100MB, the limit on the Python Index Package manager (pip). It is used to create and publish the package to pip.

But I did found that cmp.stages.eeg was missing in the package list of https://github.com/connectomicslab/connectomemapper3/blob/master/setup_pypi.py...

While this will be fixed in the next release, there still exists a solution in the meantime: to use [setup.py](https://github.com/connectomicslab/connectomemapper3/blob/master/setup.py) that list all packages properly. Having the `py37cmp-gui` environment activated, this can be done as follows:

```
$ pip uninstall connectomemapper
$ pip install git+https://github.com/connectomicslab/connectomemapper3.git@v3.0.3
```
