### 2022-04-13 92139b1
FIX: Add missing `cmp.stages.eeg` to `setup_pypi.py`
