### 2022-04-13 aaceac6
DOC: Add sphinx-copybutton to docs/requirements.txt
