+++ docs/requirements.txt
@@ -8,6 +8,7 @@ sphinx-gallery==0.10.1
 recommonmark==0.7.1
 readthedocs-sphinx-ext==2.1.4
 sphinx-autodoc-typehints==1.12.0
+sphinx-copybutton==0.5.0
 traits==6.3.2
 traitsui==7.2.0
 grabbit
