### 2022-04-13 e34ac82
DOC: Add 'sphinx_copybutton' to extensions list in docs/conf.py
