### 2022-04-13 cc4fccb
DOC: Add customization variable of sphinx-copybutton (commented for the moment) [skip ci]
