+++ docs/conf.py
@@ -366,3 +366,19 @@ apidoc_extra_args = ["--module-first",
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {"http://docs.python.org/3": None}
+
+
+# -- sphinx-copybutton extension configuration -------------------------------
+
+# copybutton_prompt_text = "myinputprompt"
+# copybutton_prompt_text = ">>> "
+# copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
+# copybutton_prompt_is_regexp = True
+# copybutton_only_copy_prompt_lines = False
+# copybutton_remove_prompts = False
+# copybutton_copy_empty_lines = False
+# copybutton_line_continuation_character = "\\"
+# copybutton_here_doc_delimiter = "EOT"
+
+# See https://sphinx-copybutton.readthedocs.io/en/latest/use.html#keep-empty-lines
+# for more details
