### 2022-04-13 804bb73
DOC: Set copybutton prompt text regex [skip ci]
