+++ docs/conf.py
@@ -370,10 +370,8 @@ intersphinx_mapping = {"http://docs.python.org/3": None}
 
 # -- sphinx-copybutton extension configuration -------------------------------
 
-# copybutton_prompt_text = "myinputprompt"
-# copybutton_prompt_text = ">>> "
-# copybutton_prompt_text = r">>> |\.\.\. |\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
-# copybutton_prompt_is_regexp = True
+copybutton_prompt_text = r">>> |\.\.\. |\$ | |(py37cmp-gui)\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
+copybutton_prompt_is_regexp = True
 # copybutton_only_copy_prompt_lines = False
 # copybutton_remove_prompts = False
 # copybutton_copy_empty_lines = False
