### 2022-04-13 e16fa9c
DOC: Simplify regex of copybutton [skip ci]
