+++ docs/conf.py
@@ -370,10 +370,10 @@ intersphinx_mapping = {"http://docs.python.org/3": None}
 
 # -- sphinx-copybutton extension configuration -------------------------------
 
-copybutton_prompt_text = r">>> |\.\.\. |\$ | |(py37cmp-gui)\$ |In \[\d*\]: | {2,5}\.\.\.: | {5,8}: "
+copybutton_prompt_text = r">>> |\.\.\. |\$ | |\(py37cmp-gui\)\$ "
 copybutton_prompt_is_regexp = True
 # copybutton_only_copy_prompt_lines = False
-# copybutton_remove_prompts = False
+copybutton_remove_prompts = True
 # copybutton_copy_empty_lines = False
 # copybutton_line_continuation_character = "\\"
 # copybutton_here_doc_delimiter = "EOT"
