### 2022-04-13 67d076d
DOC: Add notes about docker image and conda environment size and possible long time to install
