### 2022-04-13 96a299d
DOC: Review note for docker image size [skip ci]
