### 2022-04-13 f55c132
DOC: Apply suggestions from code review
Co-authored-by: Stephan Heunis <jsheunis@gmail.com>
