### 2022-04-13 e5ed46d
FIX: Rename COPYRIGHT to LICENSE in Dockerfile
