### 2022-04-14 70c83d7
Fix typo in connectome.py
occured -> occurred
