### 2022-04-15 connectomicslab/connectomemapper3#171 @jsheunis
BUG: Running `cmpbidsappmanager` just hangs, no GUI appearing
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

When running `cmpbidsappmanager`, the following appears in the terminal:

```
>> cmpbidsappmanager

Connectome Mapper v3.0.3 - BIDS App Manager
------------------------------------------------------
Copyright (C) 2009-2022, Ecole Polytechnique Federale de Lausanne (EPFL) the University Hospital Center and University of Lausanne (UNIL-CHUV), Switzerland, and Contributors, All rights reserved.
------------------------------------------------------
------------------------------------------------------
  .. INFO: Use qt for graphical backend
------------------------------------------------------

_
```

then an icon appears in my Mac's doc bar:

and after that nothing happens. Have tried multiple times, no difference. I'm expecting the GUI to appear, is that what should happen? Should I do something about the note regarding `qt`?
