### 2022-04-22 connectomicslab/connectomemapper3#171 @sebastientourbier
Hi @jsheunis,

Sorry to hear you are still having issues with the GUI...

This is indeed not the expected behaviour and a window should appear at this point.

I tried to reproduce it on my macbookpro (MacOS Catalina 10.15.7 / conda 4.11.0) (I completely uninstalled connectomemapper3 and the environment and I reinstalled it following the instructions in the docs and the workaround in #160 ), but it appears `cmpbidsappmanager` starts without any problem on my machine.

**Possible source of problem**
This made me suspecting that it might be related to Qt-based GUI on Big Sur and after searching on internet, I discovered that several Qt-based applications had a similar issue when using PyQt/PySide2 >= 5.13.2 and < 5.15.2:
- https://stackoverflow.com/questions/64833558/apps-not-popping-up-on-macos-big-sur-11-0-1
- https://bugreports.qt.io/browse/QTBUG-87014
- https://stackoverflow.com/questions/64818879/is-there-any-solution-regarding-to-pyqt-library-doesnt-work-in-mac-os-big-sur/64856281
- https://keremkoseoglu.wordpress.com/2020/11/14/pyqt5-problem-on-macos-big-sur-solved/
- https://tickets.metabrainz.org/browse/PICARD-1858

In our case, PyQt/PySide2 5.13.2 is indeed installed in our `py37cmp-gui` environment:
https://github.com/connectomicslab/connectomemapper3/blob/bb40f8e78020814579da3ea8fc29e63437ad238c/conda/environment_macosx.yml#L19

In this sense, the closest issue would correspond to the last enumerated one (https://tickets.metabrainz.org/browse/PICARD-1858) that was fixed here: https://github.com/metabrainz/picard/pull/1633#issue-708781859. Briefly, it consisted in downgrading PyQt from 5.13.2 to 5.13.1.

**Possible solution**
Please could you try to modify the `environment_macosx.yml` by changing the line ` - pyside2=5.13.2` to  `- pyside2=5.13.1` and reinstall the environment as follows?
```bash
conda deactivate # make sure the py37cmp-gui environment is not activated
conda env remove -n py37cmp-gui
conda env create -f /path/to/edited/environment_macosx.yml
```

As regards

> Should I do something about the note regarding qt?

I would say here no.

The tool used for graphical backend of the TraitsUI elements, which is selected by setting the environment variables `ETS_TOOLKIT` and `QT_API`:

https://github.com/connectomicslab/connectomemapper3/blob/bb40f8e78020814579da3ea8fc29e63437ad238c/cmp/cli/cmpbidsappmanager.py#L14-L16

and the message `INFO: Use qt for graphical backend` is just used to confirm that `ETS_TOOLKIT` has been set to `qt`:

https://github.com/connectomicslab/connectomemapper3/blob/bb40f8e78020814579da3ea8fc29e63437ad238c/cmp/cli/cmpbidsappmanager.py#L31

Let me know if this works or if you have any further questions.

Best
