### 2022-04-30 90e08d4
DOC: Add first draft of summary in JOSS paper as suggested by reviewer [skip ci]
