+++ paper/paper.md
@@ -75,6 +75,25 @@ bibliography: paper.bib
 # aas-journal: Astrophysical Journal <- The name of the AAS journal.
 ---
 
+# Summary
+
+Connectome Mapper 3 (CMP3) is a processing pipeline software for multi-scale multi-modal connectome mapping
+  of the human brain.
+It provides to researchers an unique workflow, implemented in accordance to the BIDS App framework,
+  that leverages a number of widely adopted software tools to map a connectome representation of the
+  structural and functional brain systems at multiple scales.
+The workflow can take any structural / diffusion / resting-state functional MRI dataset structured accordingly
+  To the BIDS standard, and is intended to be run via its Docker and Singularity containerized versions.
+To improve accessibility, Connectome Mapper 3 comes with an interactive graphical user interface, which supports
+  the user in all the steps involved in the configuration of the pipelines, the configuration and execution of
+  the BIDS App, and the control of the output quality.
+The tool is written in Python, available from the Python Package Index (PyPI), and the container images
+  are available from DockerHub and SylabsIO.
+Having been successfully employed in a number of research papers, and being currently extended to
+  electroencephalography, we believe that CMP3 can offer a solid multi-modal tool for the investigation of
+  brain network function and organization at specific scales, as well as a map that links different spatial
+  and temporal scales.
+
 # Statement of Need
 
 The field of Magnetic Resonance Imaging (MRI) Connectomics has rapidly expanded since its advent 
@@ -128,7 +147,7 @@ Initiated during OHBM BrainHack 2020 ([https://github.com/ohbm/hackathon2020/iss
   CMP3 is being extended to EEG.
 This manuscript summarizes CMP3 along with different design characteristics and principles.
 
-# Summary
+# Design and functionality overview
 
 ## A flexible and interoperable workflow for multi-modal human connectome mapping
 
