### 2022-04-30 a450ca0
DOC: Edit summary and rename section "Mention" to "Community impact" [skip ci]
