### 2022-05-08 connectomicslab/connectomemapper3#171 @jsheunis
I did as you suggested, but still had the same issue.

After removing the existing conda environment, and installing a new one from the edited `environment_macosx.yml` file, I ran `pip list`. This included the following:

```
PyQt5                         5.12.3
PyQt5_sip                     4.19.18
PyQtChart                     5.12
PyQtWebEngine                 5.12.1
```

I'm not deeply familiar with the connection between `pyside2` and `PyQt`, but I'm guessing after setting `pyside2=5.12.1` in the environment file, we'd want `PyQt5` to have the same version number? If so, this doesn't seem to be happening during the conda environment creation.

I tried uninstalling `PyQt5` and reinstalling it with the specific version:

```
pip uninstall PyQt5
pip install PyQt5==5.12.1
pip list
>>
PyQt5                         5.12.1
PyQt5_sip                     4.19.18
PyQtChart                     5.12
PyQtWebEngine                 5.12.1
```

which executed successfully, but the initial issue still remained. Any other ideas?

If this persists, I'm not sure that this issue should necessarily hold up the JOSS review if it's just an edge case.
