### 2022-05-08 connectomicslab/connectomemapper3#177 @jsheunis
MAINT/DOC: Possibly upgrade datalad version
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

The latest release of datalad (currently) is `0.16.2`, which contains significants improvements (including performance related). You can have a look at the [changelog](https://github.com/datalad/datalad/blob/master/CHANGELOG.md#0160-fr-apr-8-2022-----spring-cleaning) for specifics. So I would suggested upgrading your dependency version to `>=0.16.0`, although I realise that would require some testing and also documentation updates, and might not be high priority.
