### 2022-05-08 connectomicslab/connectomemapper3#177 @jsheunis
Another note: the environment configs specify the datalad version at `0.15.4`, while the [docs](https://connectome-mapper-3.readthedocs.io/en/latest/datalad.html) state that the datalad workflow was tested with `0.14.0`. It might be better to remove this discrepancy by ensuring that the workflow works with whatever the installed datalad version will be, and updating the docs accordingly.
