### 2022-05-08 connectomicslab/connectomemapper3#178 @jsheunis
BUG: Installing tutorial dataset via ssh fails if user does not have an ssh key setup
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

I followed the [analysis tutorial](https://connectome-mapper-3.readthedocs.io/en/latest/notebooks/analysis_tutorial.html) and received an error when installing the dataset:
```
vepcon_data = dl.install(
    path=dataset_dir,
    source="git@github.com:OpenNeuroDatasets/ds003505.git"
)
>>>
Cloning dataset to Dataset(/Users/jsheunis/Downloads/ds003505_demo)
[INFO] Attempting to clone from git@github.com:OpenNeuroDatasets/ds003505.git to /Users/jsheunis/Downloads/ds003505_demo
git@github.com: Permission denied (publickey).
```
Since the source url is the one for cloning via ssh, and I didn't have an ssh key setup for this user, this step failed. Replacing the URL with the htttps equivalent succeeded:

```
vepcon_data = dl.install(
    path=dataset_dir,
    source="https://github.com/OpenNeuroDatasets/ds003505.git"
)
>>>
[INFO] Cloning dataset to Dataset(/Users/jsheunis/Downloads/ds003505_demo)
[INFO] Attempting to clone from https://github.com/OpenNeuroDatasets/ds003505.git to /Users/jsheunis/Downloads/ds003505_demo
[INFO] Start enumerating objects
[INFO] Start receiving objects
[INFO] Start resolving deltas
[INFO] Completed clone attempts for Dataset(/Users/jsheunis/Downloads/ds003505_demo)
[INFO] scanning for annexed files (this may take some time)
[INFO] Remote origin not usable by git-annex; setting annex-ignore
[INFO] https://github.com/OpenNeuroDatasets/ds003505.git/config download failed: Not Found
[INFO] access to 1 dataset sibling s3-PRIVATE not auto-enabled, enable with:
| 		datalad siblings -d "/Users/jsheunis/Downloads/ds003505_demo" enable -s s3-PRIVATE
```

(the enable warning can be ignored, the dataset still clones successfully).

I think it's safer to assume people won't have ssh keys setup and therefore to rather use https.
