### 2022-05-08 connectomicslab/connectomemapper3#179 @jsheunis
BUG: File not found `cmtklib/data/parcellation/lausanne2018/rh.lausanne2018.scale3.annot` in the analysis tutorial
_Issue related to JOSS review: https://github.com/openjournals/joss-reviews/issues/4248_

In the analysis tutorial, in the last code block:

```
%%time
# Plot
plot_lausanne2018_surface_ctx(G_fd.U[:,1], scale='scale3', save_fig=True)
>>>
FileNotFoundError                         Traceback (most recent call last)
<timed eval> in <module>

~/opt/miniconda3/envs/cmp3-tutorial/lib/python3.7/site-packages/cmtklib/data/parcellation/viz.py in plot_lausanne2018_surface_ctx(roi_values, scale, cmap, save_fig, output_dir, filename, fmt)
     63
     64     # Read annot files
---> 65     annot_right = nib.freesurfer.read_annot(annots[0])
     66     annot_left = nib.freesurfer.read_annot(annots[1])
     67

~/opt/miniconda3/envs/cmp3-tutorial/lib/python3.7/site-packages/nibabel/freesurfer/io.py in read_annot(filepath, orig_ids)
    357         The names of the labels. The length of the list is n_labels.
    358     """
--> 359     with open(filepath, "rb") as fobj:
    360         dt = _ANNOT_DT
    361

FileNotFoundError: [Errno 2] No such file or directory: '/Users/jsheunis/opt/miniconda3/envs/cmp3-tutorial/lib/python3.7/site-packages/cmtklib/data/parcellation/lausanne2018/rh.lausanne2018.scale3.annot'
```

This is what the file tree looks like:

```bash
> cd /Users/jsheunis/opt/miniconda3/envs/cmp3-tutorial/lib/python3.7/site-packages/cmtklib/data/
> tree
.
├── parcellation
│   ├── __init__.py
│   ├── __pycache__
│   │   ├── __init__.cpython-37.pyc
│   │   ├── util.cpython-37.pyc
│   │   └── viz.cpython-37.pyc
│   ├── lausanne2018
│   │   └── mni-space
│   │       ├── atlas-L2018_res-scale1_coords.csv
│   │       ├── atlas-L2018_res-scale2_coords.csv
│   │       ├── atlas-L2018_res-scale3_coords.csv
│   │       ├── atlas-L2018_res-scale4_coords.csv
│   │       └── atlas-L2018_res-scale5_coords.csv
│   ├── util.py
│   └── viz.py
└── report
    └── carbonfootprint
        ├── css
        │   └── bootstrap-combined-no-icons.2.3.2.min.css
        └── js
            └── jquery.3.3.1.min.js
```
