### 2022-05-09 connectomicslab/connectomemapper3#178 @sebastientourbier
Thank you @jsheunis for reporting this!
Yes, indeed using https would be safer. I am going to take care of making the appropriate change in the notebook.
Also, I am going to check the notebook for the EEG pipeline. It might not use https too.
