+++ paper/paper.md
@@ -77,24 +77,24 @@ bibliography: paper.bib
 
 # Summary
 
-Connectome Mapper 3 (CMP3) is a processing pipeline software for multi-scale multi-modal connectome mapping
+Connectome Mapper 3 (CMP3) is a processing pipeline software, written in Python 3, for multi-scale multi-modal connectome mapping
   of the human brain.
-It provides to researchers an unique workflow, implemented in accordance to the BIDS App framework
+It provides researchers with an unique workflow, implemented in accordance to the BIDS App framework
   [@GorgolewskiBIDSMethods:2017], that leverages a number of widely adopted software tools to map a connectome
   representation of the structural and functional brain systems at multiple scales.
 The workflow can take any structural / diffusion / resting-state functional MRI dataset structured accordingly
   to the BIDS standard [@GorgolewskiTheExperiments:2016], and is intended to be run via its Docker and
   Singularity containerized versions.
-To improve accessibility, Connectome Mapper 3 comes with an interactive graphical user interface, which supports
+To improve its accessibility, CMP3 comes with a graphical user interface, which supports
   the user in all the steps involved in the configuration of the pipelines, the configuration and execution of
   the BIDS App, and the control of the output quality.
-The tool is written in Python, available from the [Python Package Index (PyPI)](https://pypi.org/project/connectomemapper/),
-  and the container images are available from [DockerHub](https://hub.docker.com/r/sebastientourbier/connectomemapper-bidsapp)
-  and [Sylabs Cloud](https://cloud.sylabs.io/library/connectomicslab/default/connectomemapper-bidsapp).
-Having been successfully employed in a number of research papers, and being currently extended to
-  electroencephalography, we believe that CMP3 can offer a solid multi-modal tool for the investigation of
+CMP3 has been successfully employed in a number of research papers, and is now being currently extended to
+  electroencephalography, to consistute a solid multi-modal framefork for the investigation of
   brain network function and organization at specific scales, as well as a map that links different spatial
   and temporal scales.
+CMP3 is  available from the [Python Package Index (PyPI)](https://pypi.org/project/connectomemapper/),
+  and the container images available from [DockerHub](https://hub.docker.com/r/sebastientourbier/connectomemapper-bidsapp)
+  and [Sylabs Cloud](https://cloud.sylabs.io/library/connectomicslab/default/connectomemapper-bidsapp).
 
 # Statement of Need
 
