### 2022-05-09 connectomicslab/connectomemapper3#171 @sebastientourbier
That's what I would have also done...

Another idea could be to update PyQt5 to the latest version (as it seems it was fixed in version >= 5.15.2), and choose pyqt5 as backend directly.

1. Reinstall the latest version of PyQt5:
```
$ pip uninstall PyQt5
$ pip install PyQt5==5.15.6
```

2. Uncomment line 15 and comment line 16 in the file `cmpbidsappmanager.py`:

   https://github.com/connectomicslab/connectomemapper3/blob/bb40f8e78020814579da3ea8fc29e63437ad238c/cmp/cli/cmpbidsappmanager.py#L14-L16

    It should be found in the `.../miniconda3/envs/py37cmp-gui/lib/python3.7/site-packages/cmp/cli` folder. This will tell `traitsui` to use directly PyQt5 as backend.

Hope this will work 🤞
