### 2022-05-10 connectomicslab/connectomemapper3#177 @sebastientourbier
The changelog of the new version 0.16 of datalad is indeed very big! I was not aware of it. Thank you very much :)
It should definitely need to be updated and I am adding this task with a `medium-priority` label to not delay higher priorities.

The workflow presented in the tutorial on "how adopting datalad for collaboration" should  indeed use commands that are not specific to a version.
The version of Datalad and its extensions referred here corresponds to the version I used the last time I edited and tested the tutorial. I just wanted to keep track of it.
But if this is confusing, I could either specify it better, or remove completely any mention. What do you think?
