+++ paper/paper.md
@@ -79,17 +79,17 @@ bibliography: paper.bib
 
 Connectome Mapper 3 (CMP3) is an open-source processing pipeline software, written in Python 3, for multi-scale multi-modal connectome mapping
   of the human brain.
-It provides researchers with an unique workflow, implemented in accordance to the BIDS App framework
+It provides researchers with a unique workflow, implemented in accordance with the BIDS App framework
   [@GorgolewskiBIDSMethods:2017], that leverages a number of widely adopted software tools to map a connectome
   representation of the structural and functional brain systems at multiple scales.
 The workflow can take any structural / diffusion / resting-state functional MRI dataset structured accordingly
   to the BIDS standard [@GorgolewskiTheExperiments:2016], and is intended to be run via its Docker and
-  Singularity containerized versions.
+  Singularity/Apptainer containerized versions.
 To improve its accessibility, CMP3 comes with a graphical user interface, which supports
   the user in all the steps involved in the configuration of the pipelines, the configuration and execution of
   the BIDS App, and the control of the output quality.
 CMP3 has been successfully employed in a number of research papers, and is now being currently extended to
-  electroencephalography, to offer a solid multi-modal framefork to the community for the investigation of
+  electroencephalography, to offer a solid multi-modal framework to the community for the investigation of
   brain network function and organization at specific scales, as well as a map that links different spatial
   and temporal scales.
 CMP3 is available from the [Python Package Index (PyPI)](https://pypi.org/project/connectomemapper/),
