### 2022-05-17 966e7cb
FIX: Correct loop over views to render in plot_lausanne2018_surface_ctx() [skip ci]
