### 2022-05-18 714fe9e
VERSION: Update version in cmp/info.py to 3.0.4
