+++ cmp/info.py
@@ -7,9 +7,9 @@
 
 _version_major = 3
 _version_minor = 0
-_version_micro = 3
+_version_micro = 4
 _version_extra = ""
-__release_date__ = "21.02.2022"
+__release_date__ = "21.05.2022"
 
 __minor_version__ = "%s.%s" % (_version_major, _version_minor)
 
