+++ conda/environment.yml
@@ -15,7 +15,7 @@ dependencies:
     - traitsui=7.2.0
     - graphviz=2.50.0
     - configparser=5.2.0
-    - nipype=1.7.0
+    - nipype=1.8.0
     - git-annex=8.20211123
     - pyside2=5.13.2
     - indexed_gzip=1.6.4
