+++ .circleci/config.yml
@@ -389,7 +389,7 @@ jobs:
       image: ubuntu-2004:202201-02
       # OLD: Ubuntu 16.04, docker 18.09.3, docker-compose 1.23.1
       # image: ubuntu-1604:201903-01
-      docker_layer_caching: false
+      docker_layer_caching: true
     resource_class: medium
     working_directory: /home/circleci/src/connectomemapper3
     steps:
@@ -442,8 +442,13 @@ jobs:
             fi
             # Build CMP BIDS App docker image
             cd /home/circleci/src/connectomemapper3
-            ls -la .
-            DOCKER_BUILDKIT=1 docker build --progress=plain --squash --rm=true --build-arg BUILD_DATE=`date -u +"%Y-%m-%dT%H:%M:%SZ"` --build-arg VCS_REF=`git rev-parse --short HEAD` --build-arg VERSION="${CIRCLE_TAG:-$THISVERSION}" -t sebastientourbier/connectomemapper3 /home/circleci/src/connectomemapper3 \
+            export DOCKER_BUILDKIT=1
+            docker build --progress=plain \
+              --squash --rm=true \
+              --build-arg BUILD_DATE=`date -u +"%Y-%m-%dT%H:%M:%SZ"` \
+              --build-arg VCS_REF=`git rev-parse --short HEAD` \
+              --build-arg VERSION="${CIRCLE_TAG:-$THISVERSION}" \
+              -t sebastientourbier/connectomemapper3 /home/circleci/src/connectomemapper3
       #- run:
       #    name: Prune Docker cache (temporary)
       #    command: docker system prune -f
