### 2022-05-18 f49030a
DOC: Add changes for 3.0.4
