### 2022-05-20 cc9fcd8
CI: Change docker tmp dir prior to restarting the docker daemon
