### 2022-05-21 c67c974
CI: Roll back to not using docker layer caching because of lack of space
