### 2022-06-02 2d2da20
DOC: Apply suggestions from @mschoettner [skip ci]
Co-authored-by: Mikkel Schöttner <48212821+mschoettner@users.noreply.github.com>
