### 2022-06-02 6df2106
DOC: Apply suggestions from code review of @jwirsich [skip ci]
