### 2022-06-02 1467af9
DOC: Move paper sections for better rendering [skip ci]
