+++ paper/paper.md
@@ -236,35 +236,6 @@ All connectome files employ a common naming convention, based on the current
   for more details).
 \label{fig:parc}](Lausanne2018_parcellation_diagram.png)
 
-## Outputs ready to be reused in the BIDS ecosystem
-
-CMP3 outputs follow the BIDS Derivatives specifications wherever possible,
-  which facilitates the sharing of the derivatives in the BIDS App ecosystem,
-  and allows the user to easily retrieve any of the files generated by CMP3
-  with tools of the BIDS ecosystem such as pybids [@Yarkoni:2019].
-It introduces a new BIDS entity ``atlas-<atlas_label>`` (See [proposal](https://github.com/bids-standard/bids-specification/pull/997))
-  that is used in combination with the ``res-<atlas_scale>`` entity to distinguish imaging and network data derived
-  from different parcellation atlases and scales (\autoref{fig:parc}).
-While the BIDS-Derivatives extension to organize network data
-  (See [BEP017](https://docs.google.com/document/d/1ugBdUF6dhElXdj3u9vw0iWjE6f_Bibsro3ah7sRV0GA/edit#heading=h.mqkmyp254xh6))
-  is being developed, in which we are actively participating, structural and functional connectome files
-  derived from the different imaging modalities are saved in multiple formats following the convention shown in \autoref{fig:parc}.
-All connectomes are saved by default as graph edge lists in ``.tsv`` files, that can
-  be directly analyzed using \href{https://networkx.org/documentation/stable/tutorial.html}{NetworkX} [@Hagberg:2008],
-  a Python library which offers many algorithms and tools to explore graphs and compute local and global network properties.
-Connectivity matrices can be exported to MATLAB as MAT-files can be fed to the
-  \href{www.brain-connectivity-toolbox.net}{Brain Connectivity Toolbox} [@Rubinov:2010], which is a powerful
-  toolbox containing a large selection of network measures for the characterization of brain
-  connectivity datasets.
-Finally, connectomes can be saved in GraphML format to interface with a lot of general purpose
-  software packages for graph analysis such as \href{www.cytoscape.org}{Cytoscape} [@Shannon:2003] [@Gustavsen:2019]
-  or \href{www.gephi.org}{Gephi} [@Bastian:2009].
-Structuring outputs as BIDS Derivatives and saving them in a range of file formats
-  thus has a lot of advantages.
-Not only does it ensure that the connectome files can be opened by the most popular
-  software packages used in this field to perform complex network analyses, but it
-  also eases the reuse of all outputs in the BIDS ecosystem.
-
 ## A focus on accessibility and versatility
 
 CMP3 takes advantage of the Traits/TraitsUI framework
@@ -296,11 +267,40 @@ It offers them the possibility to tune and save all the parameters in configurat
   can then be employed for running the BIDS App either with the Docker or Singularity software container engine directly,
   or with the two [lightweight Docker and Singularity wrappers](https://connectome-mapper-3.readthedocs.io/en/latest/usage.html#with-the-wrappers).
 
+## Outputs ready to be reused in the BIDS ecosystem
+
+CMP3 outputs follow the BIDS Derivatives specifications wherever possible,
+  which facilitates the sharing of the derivatives in the BIDS App ecosystem,
+  and allows the user to easily retrieve any of the files generated by CMP3
+  with tools of the BIDS ecosystem such as pybids [@Yarkoni:2019].
+It introduces a new BIDS entity ``atlas-<atlas_label>`` (See [proposal](https://github.com/bids-standard/bids-specification/pull/997))
+  that is used in combination with the ``res-<atlas_scale>`` entity to distinguish imaging and network data derived
+  from different parcellation atlases and scales (\autoref{fig:parc}).
[…]
