### 2022-06-09 connectomicslab/connectomemapper3#171 @roryjpiper
Hi @sebastientourbier - thanks for this. This change to pyqt5 now causes a BIDS App Manager to pop up ....but it is completely blank.

This is the error message, please can you help further?:

(py37cmp-gui) rorypiper@eduroam-int-dhcp-97-64-214 software % cmpbidsappmanager
objc[55865]: Class QMacAutoReleasePoolTracker is implemented in both /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/python3.7/site-packages/PyQt5/Qt5/lib/QtCore.framework/Versions/5/QtCore (0x177bee198) and /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/libQt5Core.5.12.9.dylib (0x189671528). One of the two will be used. Which one is undefined.
objc[55865]: Class QT_ROOT_LEVEL_POOL__THESE_OBJECTS_WILL_BE_RELEASED_WHEN_QAPP_GOES_OUT_OF_SCOPE is implemented in both /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/python3.7/site-packages/PyQt5/Qt5/lib/QtCore.framework/Versions/5/QtCore (0x177bee210) and /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/libQt5Core.5.12.9.dylib (0x1896715a0). One of the two will be used. Which one is undefined.
objc[55865]: Class KeyValueObserver is implemented in both /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/python3.7/site-packages/PyQt5/Qt5/lib/QtCore.framework/Versions/5/QtCore (0x177bee238) and /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/libQt5Core.5.12.9.dylib (0x1896715c8). One of the two will be used. Which one is undefined.
objc[55865]: Class RunLoopModeTracker is implemented in both /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/python3.7/site-packages/PyQt5/Qt5/lib/QtCore.framework/Versions/5/QtCore (0x177bee288) and /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/libQt5Core.5.12.9.dylib (0x189671618). One of the two will be used. Which one is undefined.
objc[55865]: Class QCocoaPageLayoutDelegate is implemented in both /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/python3.7/site-packages/PyQt5/Qt5/lib/QtPrintSupport.framework/Versions/5/QtPrintSupport (0x177e01540) and /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/libQt5PrintSupport.5.12.9.dylib (0x18b9593b0). One of the two will be used. Which one is undefined.
objc[55865]: Class QCocoaPrintPanelDelegate is implemented in both /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/python3.7/site-packages/PyQt5/Qt5/lib/QtPrintSupport.framework/Versions/5/QtPrintSupport (0x177e015b8) and /Users/rorypiper/opt/anaconda3/envs/py37cmp-gui/lib/libQt5PrintSupport.5.12.9.dylib (0x18b959428). One of the two will be used. Which one is undefined.

Connectome Mapper v3.0.3 - BIDS App Manager
------------------------------------------------------
Copyright (C) 2009-2022, Ecole Polytechnique Federale de Lausanne (EPFL) the University Hospital Center and University of Lausanne (UNIL-CHUV), Switzerland, and Contributors, All rights reserved.
------------------------------------------------------
------------------------------------------------------
  .. INFO: Use qt for graphical backend
------------------------------------------------------
