### 2022-06-09 connectomicslab/connectomemapper3#171 @roryjpiper
Scratch that. You have solved this already - here: https://github.com/connectomicslab/connectomemapper3/issues/61

:)

Your fix on May 9th worked BTW! Thanks!
