### 2022-06-11 connectomicslab/connectomemapper3#171 @sebastientourbier
@jsheunis, did you have the chance to give a try on this:

> That's what I would have also done...
>
> Another idea could be to update PyQt5 to the latest version (as it seems it was fixed in version >= 5.15.2), and choose pyqt5 as backend directly.
>
>     1. To reinstall the latest version of PyQt5:
>
>
> ```
> $ pip uninstall PyQt5
> $ pip install PyQt5==5.15.6
> ```
>
>     2. To choos PyQt5 directly, you would need to uncomment line 15 and comment line 16 in the file `cmpbidsappmanager.py`:

 https://github.com/connectomicslab/connectomemapper3/blob/bb40f8e78020814579da3ea8fc29e63437ad238c/cmp/cli/cmpbidsappmanager.py#L14-L16

>
>        that should be found in the `.../miniconda3/envs/py37cmp-gui/lib/python3.7/site-packages/cmp/cli` folder. This will tell `traitsui` to use directly PyQt5 as backend.
>
>
> Hope this will work 🤞

?

If this solution works, I could plan to integrate it directly in the upcoming release.
