### 2022-06-13 connectomicslab/connectomemapper3#171 @jsheunis
This solved the issue, thanks!
