### 2022-06-14 70036fc
UPD: Refactor YAMLs for py37cmp-gui environment to use latest Qt5 version
