+++ conda/environment_macosx.yml
@@ -4,27 +4,24 @@ channels:
     - conda-forge
     - aramislab
     - mrtrix3
-    - anaconda
 
 dependencies:
     - python=3.7
     - pip=21.3.1
-    - numpy=1.21.5
-    - matplotlib=3.5.1
-    - traits=6.3.2
-    - traitsui=7.2.0
     - graphviz=2.50.0
-    - configparser=5.2.0
-    - nipype=1.7.0
-    - pyside2=5.13.2
-    - indexed_gzip=1.6.4
     - dipy=1.3.0
-    - cvxpy-base=1.1.18
-    - cvxpy=1.1.18
     - fsleyes=1.3.3
+    - indexed_gzip=1.6.4
     - mrtrix3=3.0.3
 
     - pip:
+          - numpy==1.21.6
+          - matplotlib==3.5.2
+          - configparser==5.2.0
+          - nipype==1.8.0
+          - traits==6.3.2
+          - traitsui==7.2.0
+          - PyQt5==5.15.6
           - duecredit==0.9.1
           - obspy==1.2.2
           - mne==0.24.1
