### 2022-06-14 9f19e6d
UPD: Change PySide2 to PyQt5 graphical backend for traitsui
