### 2022-06-14 aff16e8
FIX: Print version of Qt and PyQt5 in cmpbidsappmanager
