### 2022-06-14 f1f1999
FIX: Restart docker after setting experimental mode in .circleci/config.yml
