### 2022-06-15 f3cee31
DOC: Add changes for PyQt5 <-> PySide2 [skip ci]
