### 2022-06-15 1f41153
DOC: Add note for blank window at cmpbidsappmanager launch and clarify how to load dataset [skip ci]
