### 2022-06-15 f4deef7
DOC: Improve note on blank window at launch [skip ci]
