### 2022-06-15 1f16721
DOC: Review docs to load BIDS dataset in GUI [skip ci]
