### 2022-06-15 701c445
DOC: Improve description to load dataset with cmpbidsappmanager [skip ci]
