### 2022-06-15 bfd3309
DOC: Review previous notes and instructions [skip ci]
