### 2022-06-15 cf6ef89
DOC: Improve intro version 3.0.4 in changes [skip ci]
