### 2022-06-15 0a77ae6
DOC: Fix some typos in changes [skip ci]
