### 2022-06-15 330a7a4
DOC: Change release date in cmp/info.py [skip ci]
