### 2022-06-15 f670110
PAPER: Update Zenodo citation in paper.bib for Connectome Mapper v3.0.4
