+++ paper/paper.bib
@@ -1,10 +1,20 @@
 @software{ZenodoCMP:2022,
-    author = {Tourbier, Sebastien and Aleman-Gomez, Yasser and Mullier, Emeline and Griffa, Alessandra and Bach Cuadra, Meritxell and Hagmann, Patric},
-    title = {{connectomicslab/connectomemapper3: Connectome Mapper v3.0.3}},
-    month = {2},
+    author = {Tourbier, Sebastien and Rue Queralt, Joan and
+              Glomb, Katharina and
+              Aleman-Gomez, Yasser and
+              Mullier, Emeline and
+              Griffa, Alessandra and
+              Schöttner, Mikkel and
+              Wirsich, Jonathan and
+              Tuncel, Anil and
+              Jancovic, Jakub and
+              Bach Cuadra, Meritxell and
+              Hagmann, Patric},
+    title = {{connectomicslab/connectomemapper3: Connectome Mapper v3.0.4}},
+    month = {6},
     year = {2022},
     publisher = {Zenodo},
-    version = {v3.0.3},
+    version = {v3.0.4},
     doi = {10.5281/zenodo.3475969},
     url = {https://doi.org/10.5281/zenodo.3475969}
 }
