### 2022-06-16 #4248 @sebastientourbier
Dear all,

I would like to let you know that the release of a new version (3.0.4) that includes all changes is now available.

The new paper is now up-to-date with the new high-level summary on the `master` branch.

You can consult a summary of the changes at the following URL: https://github.com/connectomicslab/connectomemapper3/releases/tag/v3.0.4

Let me know if I need to do something else.

Best regards
