### 2022-06-23 #4248 @sebastientourbier
@osorensen Here is the information you need related to the latest release:

* Tag release: [` v3.0.4 `](https://github.com/connectomicslab/connectomemapper3/releases/tag/v3.0.4)
* DOI of the Zenodo record: [10.5281/zenodo.6645256](http://doi.org/10.5281/zenodo.6645256)

I also changed the title of the record to match exactly the title of the paper. I also checked the list of authors/orcid and they are the same.

Let me know if you need anything else.
