### 2022-06-23 8423ab3
FIxed a doi causing an issue for JOSS paper
Related to [this review](https://github.com/openjournals/joss-reviews/issues/4248).
