+++ paper/paper.bib
@@ -2520,7 +2520,7 @@
   author = {Pascucci, D. and Tourbier, S. and Rue-Queralt, J. and Carboni, M. and Hagmann, P. and Plomp, G.},
   title = {"VEPCON: Source imaging of high-density visual evoked potentials with multi-scale brain parcellations and connectomes"},
   year = {2021},
-  doi = {doi:10.18112/openneuro.ds003505.v1.0.3},
+  doi = {10.18112/openneuro.ds003505.v1.0.3},
   publisher = {OpenNeuro}
 }
 
