### 2022-06-23 b5f5d34
Update paper.bib
I think my last PR was in the wrong branch. This should fix the doi issue in the [JOSS review](https://github.com/openjournals/joss-reviews/issues/4248).
