### 2022-06-23 aff50d0
Suggested rewrite
Related to JOSS review https://github.com/openjournals/joss-reviews/issues/4248
