### 2022-06-23 dfe0189
updated formatting of references
