+++ paper/paper.md
@@ -122,7 +122,7 @@ The size and organization of these datasets as well as the increasing complexity
   reproducibility and transparency of the analysis.
 The last ten years have indeed witnessed a number of technical advances and a trend towards the adoption
   of open research practices such as *open data* and *open methods*
-  [@Nichols:2017] [@EglenTowardNeuroscience:2017] [@Kennedy2019] [@Poldrack2019ComputationalNeuroimaging].
+  [@EglenTowardNeuroscience:2017;@Kennedy2019;@Nichols:2017;@Poldrack2019ComputationalNeuroimaging].
 This has led in particular to the creation of a community standard for dataset organization, known as the
   Brain Imaging Data Structure (BIDS), designed to ease the practice of sharing raw neuroimaging data [@GorgolewskiTheExperiments:2016]. 
 Combined with advances in software virtualization, BIDS has enabled the creation of the BIDS Apps
