### 2022-06-23 connectomicslab/connectomemapper3#193 @osorensen
Spelling error in Figure 1
This relates to JOSS review https://github.com/openjournals/joss-reviews/issues/4248

Please rewrite from "distorsion correction" to "distortion correction" in the two places encircled. I also suggest rewriting "configurations file" to "configuration file" in each of the three columns on top of the figure.

![image](https://user-images.githubusercontent.com/21175639/175384594-4e04361c-a7b2-4ccf-a847-7d4dea895da7.png)
