+++ paper/paper.md
@@ -288,7 +288,7 @@ Connectivity matrices can be exported to MATLAB as MAT-files can be fed to the
   toolbox containing a large selection of network measures for the characterization of brain
   connectivity datasets.
 Finally, connectomes can be saved in GraphML format to interface with a lot of general purpose
-  software packages for graph analysis such as \href{www.cytoscape.org}{Cytoscape} [@Shannon:2003] [@Gustavsen:2019]
+  software packages for graph analysis such as \href{www.cytoscape.org}{Cytoscape} [@Gustavsen:2019;@Shannon:2003]
   or \href{www.gephi.org}{Gephi} [@Bastian:2009].
 Structuring outputs as BIDS Derivatives and saving them in a range of file formats
   thus has a lot of advantages.
