### 2022-06-23 a31e02e
fixed inconsistent Python spelling
