### 2022-06-23 connectomicslab/connectomemapper3#193 @osorensen
Sorry, distorsion seems to be a correct spelling.
