### 2022-06-23 #4248 @osorensen
@sebastientourbier can you please try to remove the `isbn` field from `paper.bib`. It seems like two papers have such a field, which might cause the issue above. Please let me know when it's done.
