### 2022-06-23 #4248 @osorensen
Here are the lines: https://github.com/connectomicslab/connectomemapper3/search?q=isbn&type=
