### 2022-06-24 connectomicslab/connectomemapper3#193 @sebastientourbier
Still it seems distortion is more correct as the root in English is the word distort. I am currently making the changes in the figure.
