### 2022-06-24 connectomicslab/connectomemapper3#193 @osorensen
Excellent!
