### 2022-06-24 644937c
JOSS REV: Remove isbn in entries of paper.bib
