+++ paper/paper.bib
@@ -42,7 +42,6 @@
     number = {3},
     pages = {299--303},
     volume = {20},
-    isbn = {9788578110796},
     doi = {10.1038/nn.4500},
     issn = {15461726},
     pmid = {25246403},
@@ -79,7 +78,6 @@
     number = {1},
     pages = {111--116},
     volume = {16},
-    isbn = {1532-4435},
     doi = {10.1038/s41592-018-0235-4},
     issn = {1548-7091},
     pmid = {30532080},
@@ -191,7 +189,6 @@
     pages = {e1005209},
     volume = {13},
     url = {https://dx.plos.org/10.1371/journal.pcbi.1005209},
-    isbn = {1111111111},
     doi = {10.1371/journal.pcbi.1005209},
     issn = {1553-7358}
 }
@@ -264,7 +261,6 @@
     author = {Hagmann, Patric},
     volume = {3230},
     school = {EPFL},
-    isbn = {9783642312083},
     doi = {10.5075/epfl-thesis-3230}
 }
 
