### 2022-06-24 #4248 @sebastientourbier
@osorensen All isbn fields have been removed 👍
