### 2022-06-25 #4248 @Kevin-Mattheus-Moerman
@sebastientourbier
- I checked the archive meta-data and all looks good. :+1:
- I checked the version tag for the review and archive and verified they match  :+1:
- I proofread the paper and have the following comment:
- [ ] Introduce `Magnetic Resonance Imaging (MRI)` in the summary where MRI is first used.
