### 2022-06-27 588233f
REV: Introduce MRI and BIDS in the paper summary
