### 2022-06-27 89aac62
REV: Remove redefinition of CMP3 acronym in paper [skip ci]
