### 2022-06-27 8bbe622
REV: Define BIDS the very first time it is used in the paper
