+++ paper/paper.md
@@ -79,12 +79,12 @@ bibliography: paper.bib
 
 Connectome Mapper 3 (CMP3) is an open-source processing pipeline software, written in Python 3,
   for multi-scale multi-modal connectome mapping of the human brain.
-It provides researchers with a unique workflow, implemented in accordance with the BIDS App framework
-  [@GorgolewskiBIDSMethods:2017], that leverages a number of widely adopted software tools to map a graph
+It provides researchers with a unique workflow, implemented in accordance with the Brain Imaging Data Structure (BIDS)
+  App framework [@GorgolewskiBIDSMethods:2017], that leverages a number of widely adopted software tools to map a graph
   representation of the structural and functional connections between brain regions, the connectome,
   at multiple scales.
 The workflow can take any structural / diffusion / resting-state functional Magnetic Resonance Imaging (MRI)
-  dataset structured according to the Brain Imaging Data Structure (BIDS) standard [@GorgolewskiTheExperiments:2016],
+  dataset structured according to the BIDS standard [@GorgolewskiTheExperiments:2016],
   and is intended to be run via its Docker and Singularity/Apptainer containerized versions.
 To improve its accessibility, CMP3 comes with a graphical user interface, which supports
   the user in all the steps involved in the configuration of the pipelines, the configuration and execution of
