### 2022-06-27 de7ff0a
REV: Remove redefinition of BIDS in paper [skip ci]
