### 2022-06-27 #4248 @sebastientourbier
Dear @Kevin-Mattheus-Moerman,

Thank you very much.

I just took care to introduce MRI the very first time it is used in the paper. I also took care of defining BIDS in the summary too. Finally, I noticed that I was redefining CMP3 acronym multiple times and this has been fixed. The changes can be viewed in https://github.com/connectomicslab/connectomemapper3/pull/196.

Let me know if you need anything else.
Best
