### 2022-06-27 #4248 @sebastientourbier
This is a really great 🎉
Thank you very much all for the time you invested in the reviewing/editing process. It really helped in improving the robustness and the documentation of CMP3, as well as  the quality of the paper. 🙏
